create_issue: support assigning new issues to an epic
An API exists to assign issues to an epic.  The most convenient time to do
this is when creating a new issue, so add a new positional parameter to
create-issue before Summary.  This seems to be a logical place to put it
since the first two parameters are "project" and "type" and after that the
parameters become free-form or m:n relationships.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>